% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{aggregate_quarterly_tonnage}
\alias{aggregate_quarterly_tonnage}
\title{Aggregate daily PortWatch tonnage to quarterly per commodity}
\usage{
aggregate_quarterly_tonnage(portwatch, commodities)
}
\description{
Aggregate daily PortWatch tonnage to quarterly per commodity
}
\keyword{internal}
