% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{build_other_bucket}
\alias{build_other_bucket}
\title{Construct the "other" bucket as residual-of-totals}
\usage{
build_other_bucket(feats_named, monthly_tonnage, monthly_y, monthly_prices)
}
\description{
We don't have a per-commodity tonnage/price for "other" so we use
aggregate AUS tonnage and a trade-weighted price index -- weights are
the 2019-2023 average value share of the three named commodities,
with any residual weight (unobservable) dropped.
}
\keyword{internal}
