% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_read}
\alias{cache_read}
\title{Read a cached object; returns \code{NULL} if not present}
\usage{
cache_read(cfg, source, key)
}
\description{
Read a cached object; returns \code{NULL} if not present
}
\keyword{internal}
