% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_disr.R
\name{default_disr_rows}
\alias{default_disr_rows}
\title{Default DISR T16 row mapping: commodity -> row number(s) to SUM.}
\usage{
default_disr_rows()
}
\description{
Default DISR T16 row mapping: commodity -> row number(s) to SUM.
}
\keyword{internal}
