% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_portwatch.R
\name{portwatch_list_portids}
\alias{portwatch_list_portids}
\title{List unique portids in scope via a distinct-values query.}
\usage{
portwatch_list_portids(base, countries, cfg)
}
\description{
List unique portids in scope via a distinct-values query.
}
\keyword{internal}
