% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quarter_end}
\alias{quarter_end}
\title{Quarter end for a given date}
\usage{
quarter_end(x)
}
\arguments{
\item{x}{A Date or vector of Dates.}
}
\value{
Date(s) of the last calendar day of the containing quarter.
}
\description{
Quarter end for a given date
}
\keyword{internal}
