% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tonnage_extrapolation.R
\name{seasonal_monthly_norm}
\alias{seasonal_monthly_norm}
\title{Seasonal monthly norm from daily tonnage}
\usage{
seasonal_monthly_norm(daily)
}
\description{
Seasonal monthly norm from daily tonnage
}
\keyword{internal}
