% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitc_crosswalk.R
\name{sitc_to_commodity}
\alias{sitc_to_commodity}
\title{Reverse lookup: which of our commodities does an SITC code belong to?}
\usage{
sitc_to_commodity(sitc_codes, crosswalk)
}
\arguments{
\item{sitc_codes}{Character vector of 3-digit SITC codes.}

\item{crosswalk}{The tibble returned by \code{\link[=load_sitc_crosswalk]{load_sitc_crosswalk()}}.}
}
\value{
Character vector of commodity labels, \code{NA} for unmapped codes.
}
\description{
Reverse lookup: which of our commodities does an SITC code belong to?
}
\keyword{internal}
